requirements.txt
discord-py-slash-command==3.0.3
discord.py==1.7.3
SQLAlchemy==1.4.35
